Restructure math package
