+++ src/math/core/affines/index.js
@@ -1,0 +1,2 @@
+export * from './affine2.js'
+export * from './affine3.js'
