+++ src/math/core/basis/index.js
@@ -1,0 +1,2 @@
+export * from './basis2.js'
+export * from './basis3.js'
