+++ src/math/core/constants.js
@@ -1,5 +1,3 @@
-import { sqrt } from './math.js'
-
 export const TAU = Math.PI * 2
 export const { PI } = Math
 export const HALF_PI = Math.PI / 2
@@ -7,6 +5,6 @@ export const QUARTER_PI = Math.PI / 4
 export const DEG2RAD = Math.PI / 180
 export const RAD2DEG = 180 / Math.PI
 export const epilson = 0.0000001
-export const SQRT2 = sqrt(2)
+export const SQRT2 = Math.sqrt(2)
 export const INV_SQRT2 = 1 / Math.sqrt(2)
-export const PHI = (1 + sqrt(5)) / 2
+export const PHI = (1 + Math.sqrt(5)) / 2
