+++ src/math/core/functions/index.js
@@ -1,0 +1,6 @@
+export * from './clamp.js'
+export * from './easing.js'
+export * from './indexmap.js'
+export * from './interpolation.js'
+export * from './mathematical.js'
+export * from './noise.js'
