+++ src/math/core/index.js
@@ -1,0 +1,7 @@
+export * from './affines/index.js'
+export * from './angles/index.js'
+export * from './basis/index.js'
+export * from './functions/index.js'
+export * from './matrices/index.js'
+export * from './vectors/index.js'
+export * from './constants.js'
