+++ src/math/core/matrices/index.js
@@ -1,0 +1,3 @@
+export * from './matrix2.js'
+export * from './matrix3.js'
+export * from './matrix4.js'
