+++ src/math/core/vectors/boolean/bvector2.js
