+++ src/math/core/vectors/boolean/bvector3.js
