+++ src/math/core/vectors/boolean/bvector4.js
