+++ src/math/core/vectors/float/index.js
@@ -1,0 +1,3 @@
+export * from './vector2.js'
+export * from './vector3.js'
+export * from './vector4.js'
