+++ src/math/index.js
@@ -1,17 +1,1 @@
-export * from './math.js'
-export * from './vector2.js'
-export * from './vector3.js'
-export * from './vector4.js'
-export * from './bvector2.js'
-export * from './bvector3.js'
-export * from './bvector4.js'
-export * from './quaternion.js'
-export * from './rotary.js'
-export * from './angle.js'
-export * from './matrix2.js'
-export * from './matrix3.js'
-export * from './matrix4.js'
-export * from './affine2.js'
-export * from './affine3.js'
-export * from './interpolation.js'
-export * from './constants.js'
+export * from './core/index.js'
