--- src/math/math.js
@@ -1,264 +1,0 @@
-import { DEG2RAD, epilson, RAD2DEG } from './constants.js'
-import { Vector2 } from './vector2.js'
-import { Vector3 } from './vector3.js'
-
-/**
- * Creates a random number between the parameters.
- *
- * @param {number} [min=0] - The minimal bound of the random number.
- * @param {number} [max=1] - The maximum bound of the random number.
- * @returns {number}
- */
-export function rand(min = 0, max = 1) {
-  return Math.random() * (max - min) + min
-}
-
-/**
- * Returns the square of a number.
- *
- * @param {number} x - The number to square.
- * @returns {number}
- */
-export function sq(x) {
-  return x * x
-}
-
-/**
- * Returns the power of a number by a given exponent.
- *
- * @param {number} x - The number to power.
- * @param {number} [e=2] - The number to power by.
- * @returns {number}
- */
-export function exp(x, e = 2) {
-  return x ** e
-}
-
-/**
- * Returns the square root pf a number.
- *
- * @param {number} x - The number to root.
- * @returns {number}
- */
-export function sqrt(x) {
-  return Math.sqrt(x)
-}
-
-/**
- * @param {number} value
- */
-export function invert(value) {
-  return 1 / value
-}
-
-
-/**
- * Interpolates between two numbers by a constant t.
- *
- * @param {number} from - The minimal bound of the interpolation.
- * @param {number} to - The maximum bound of the interpolation.
- * @param {number} t - A number between 0 and 1 to interpolate 
- * by.Any other number greater than 1 or less than 0 will
- * extapolate beyond b or a respectively.
- * @returns {number}
- */
-export function lerp(from, to, t) {
-  return from + t * (to - from)
-}
-
-/**
- * Gets the interpolation constant between two numbers give a value
- * between them.
- *
- * @param {number} from - The minimal bound of the interpolation.
- * @param {number} to - The maximum bound of the interpolation.
- * @param {number} value - The value between the bounds.
- * @returns {number} A number between 0 and 1 if value respects 
- * the bounds.
- */
-export function inverseLerp(from, to, value) {
-  return (value - from) / (to - from)
-}
-
-/**
- * Rounds a given value to a given precision.
- *
- * @param {number} number - The number to round.
- * @param {number} [precision=4] - How many decimal places there should be.
- * @returns {number}
- */
-export function round(number, precision = 4) {
-  const x = 10 ** precision
-
-  return Math.round(number * x) / x
-}
-
-/**
- * Clamps a value between two numbers.
- *
- * @param {number} value - The number to clamp.
- * @param {number} min - The minimal bound of the clamped number.
- * @param {number} max - The maximum bound of the clamped number.
- * @returns {number}
- */
-export function clamp(value, min, max) {
-  if (value < min) return min
-  if (value > max) return max
-
-  return value
-}
-
-/**
- * @param {number} value 
- * @param {number} step 
- * @returns {number}
- */
-export function snap(value, step) {
-  return Math.round(value / step) * step
-}
-
-/**
- * @param {number} value 
- * @param {number} step 
- * @returns {number}
- */
-export function snapDown(value, step) {
-  return Math.floor(value / step) * step
-}
-
-/**
- * @param {number} value 
- * @param {number} step 
- * @returns {number}
- */
-export function snapUp(value, step) {
-  return Math.ceil(value / step) * step
-}
-
-/**
- * Maps a value from one range to another.
- *
- * @param {number} v
- * @param {number} x1
- * @param {number} y1
- * @param {number} x2
- * @param {number} y2
- * @returns {number}
- */
-export function remap(v, x1, y1, x2, y2) {
-  return x2 + v * (y2 - x2) / (y1 - x1)
-}
-
-/**
- * Returns a unique number given from a pair of numbers.
- * @param {number} a
- * @param {number} b
- * @returns {number}
- */
-export function naturalizePair(a, b) {
-  if (a > b) return (a + b) * (a + b + 1) / 2 + a
-
-  return (a + b) * (a + b + 1) / 2 + b
-}
-
-/**
- * Converts a degree to a radian.
- *
- * @param {number} deg - Number to convert.
- * @returns {number}
- */
-export function degToRad(deg) {
-  return deg * DEG2RAD
-}
-
-/**
- * Converts a radian to a degree.
- *
- * @param {number} rad - Number to convert.
- * @returns {number}
- */
-export function radToDeg(rad) {
-  return rad * RAD2DEG
-}
-
-/**
- * @param {number} value
- * @param {number} min
- * @param {number} max
- */
-export function wrap(value, min, max) {
-  const range = max - min
-
-  return (min + ((((value - min) % range) + range) % range))
-}
-
-/**
- * @param {number} a 
- * @param {number} b 
- * @returns {number}
- */
-export function cantorPair(a, b) {
-  return 0.5 * (a + b) * (a + b + 1) + b
-
-}
-
-/**
- * @param {number} a 
- * @param {number} b 
- * @returns {number}
- */
-export function cantorPairSigned(a, b) {
-  const x = (a >= 0.0 ? 2.0 * a : (-2.0 * a) - 1.0)
-  const y = (b >= 0.0 ? 2.0 * b : (-2.0 * b) - 1.0)
-
-  return cantorPair(x, y)
-}
-
-/**
- * @param {number} value 
- * @param {number} width 
- * @returns {Vector2}
- */
-export function mapToIndex2D(value, width) {
-  return new Vector2(value % width, Math.floor(value / width))
-}
-
-/**
- * @param {number} index
- * @param {number} width
- * @param {number} height
- * @returns {Vector3}
- */
-export function mapToIndex3D(index, width, height) {
-  const depthMax = width * height
-  const rem = index % depthMax
-  const z = Math.floor(index / depthMax)
-  const y = Math.floor(rem / width)
-  const x = rem % width
-
-  return new Vector3(x, y, z)
-}
-
-/**
- * @param {number} value
- * @returns {number}
- */
-export function smoothStep(value) {
-  return value * value * (3 - 2 * value)
-};
-
-/**
- * @param {number} value
- */
-export function smootherStep(value) {
-  return value * value * value * (value * (value * 6 - 15) + 10)
-}
-
-/**
- * @param {number} a
- * @param {number} b
- * @param {number} tolerance
- */
-export function fuzzyEqual(a, b, tolerance = epilson) {
-  return Math.abs(a - b) <= tolerance
-}
